requirements.txt
wavelink==3.4.1
edge-tts>=6.1.0
redis>=5.2.1
google-generativeai>=0.3.0
gtts
psutil
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
uvicorn
fastapi
click==8.1.8
discord.py==2.5.2
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.3
packaging==24.2
propcache==0.3.1
psycopg2-binary==2.9.10
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.1.0
requests==2.32.3
spotipy==2.25.1
SQLAlchemy==2.0.40
typing_extensions==4.13.2
urllib3==2.4.0
motor>=3.3.0
apscheduler>=3.10.0
python-dateutil
icalendar
Werkzeug==3.1.3
yarl==1.19.0
openai
requests
yt-dlp
pytz
